+++ include/simplified/system/interaction/function/pairwise/core12/_pre_2_post_for_g.h
@@ -10,13 +10,13 @@
 namespace mysimulator {
 
   void _pre_2_post_for_g_pair_core12(
-      Array<Float> const& _pre, Array<Float>& _post,
+      Array<Double> const& _pre, Array<Double>& _post,
       InteractionFuncDataState& _status, const InteractionParameter& P) {
     assert((bool)_pre);
     assert((bool)_post);
     assert((bool)P);
-    float tmd=1./_pre[PairCore12PreName::PairDistanceSQ];
-    float tmd1=tmd*tmd;
+    double tmd=1./_pre[PairCore12PreName::PairDistanceSQ];
+    double tmd1=tmd*tmd;
     tmd1*=tmd1*tmd1;
     _post[PairCore12PostName::PairCore12IvDistance12]=tmd1;
     _post[PairCore12PostName::PairCore12IvDistance14]=tmd1*tmd;
